fw: dbg: cmd to periodically read a code marker pin
Use the AUX3 pin as the code marker pin.

This was done for the experiment to measure debugger energy interference
while the debugger is reading the code marker pins
(eval/dbg-interference). However, that measurement method was superceded
by the I-V curve measurement method (eval/dbg-interference-iv). The
latter does not need this command.