fw: dbg: turn on continuous power immediately after saving Vcap
This is for two reasons:
(1) minimize energy interference: there is no functional reason to use
up the target's energy from the time debugger signals the target until
debugger receives the target's acknowledgement.
(2) minimize energy interference in case when active mode is used
to implement energy-free code regions: the ackowledgement from
the target might need to encode more information (like, whether
the UART should be enabled or not), in which case the above-mentioned
overhead increases. Also, this makes it easier to drop the
acknowledgement altogether in some active mode use cases.