fw: dbg: rfid: clear all mode bits when turning off RX decoding timer
The previous code left CM1 bit set, which did not turn off the timer.
Probably, this never manifested in a problem because the RT/TR CAL
failure clause does not seem to be ever reached when reader is
configured for WISP5 modulation (it is reached if reader is in
default M4 modulation).