fw: dbg: turn off adc result sending after turning off ADC
This is not verified, but occassionally we see adc buffer overflow
assert fail, which seems to occur only on subsequence streams.
Maybe this is because on stop the ADC keeps on runnig but the
result sending to host is already disable (the delay may be
due to intervening RFID ISRs for example).